Create ProjectMetadataService to get and save project metadata
